requirements.txt
farm-haystack[elasticsearch]==1.20.1 # warning due to "transformers==4.36.0"
farm-haystack[inference]
datasets>=1.11.0
tiktoken==0.3.2
pandas
scikit-learn>=1.3.0
torch
transformers==4.36.0
sentence_transformers
#python 3.8.10